requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
customtkinter
pillow
pypdf
packaging